+++ NEWS.md
@@ -1,3 +1,7 @@
+# rtweet ...
+
+- Added support for posting alt-text metadata with images tweeted with status updated via `post_tweet()`. (@hrbrmstr)
+
 # rtweet 0.7.0
 
 - Added paper.md as part of ROpenSci submission.